+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/amazon-bedrock
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [3a602ca]
+  - @ai-sdk/provider-utils@2.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [066206e]
+  - @ai-sdk/provider-utils@2.1.4
+
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [39e5c1f]
+  - @ai-sdk/provider-utils@2.1.3
+
 ## 1.1.2
 
 ### Patch Changes